Backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
gTTS==2.3.1
idna==3.4
pytz==2022.7.1
requests==2.28.2
SpeechRecognition==3.9.0
sqlparse==0.4.3
urllib3==1.26.14
